+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 * Fix columns in the search results.
 * Fix type not being propagated to new folders from their parents.
 * Fix default type set to folders in the root folder when they are added.
+* Improve and fix comic file size format. 
 
 ## 9.14.1